+++ Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/openrr/rrt"
 
 # Note: num-traits is public dependency.
 [dependencies]
+derive_more = { version = "0.99.17", default-features = false, features = ["display", "error"] }
 kdtree = "0.7"
 num-traits = "0.2"
 rand = "0.8"
